Extracted CollectionTask.CountModels to countModels
