+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionOperations.kt
@@ -18,9 +18,13 @@ package com.ichi2.async
 import com.ichi2.anki.StudyOptionsFragment
 import com.ichi2.libanki.Card
 import com.ichi2.libanki.Collection
+import com.ichi2.libanki.Model
 import com.ichi2.libanki.Note
+import com.ichi2.utils.JSONObject
 import net.ankiweb.rsdroid.BackendFactory
 import timber.log.Timber
+import java.util.*
+import kotlin.Comparator
 
 /**
  * This file contains functions that have been migrated from [CollectionTask]
@@ -132,4 +136,29 @@ fun updateValuesFromDeck(
         Timber.e(e, "doInBackgroundUpdateValuesFromDeck - an error occurred")
         null
     }
+}
+
+/**
+ * Returns an ArrayList of all models alphabetically ordered and the number of notes
+ * associated with each model.
+ *
+ * @return {ArrayList<JSONObject> models, ArrayList<Integer> cardCount}
+ */
+fun countModels(
+    col: Collection,
+    collectionTask: ProgressSenderAndCancelListener<Void>
+): Pair<List<Model>, ArrayList<Int>>? {
+    Timber.d("doInBackgroundLoadModels")
+    val models = col.models.all()
+    val cardCount = ArrayList<Int>()
+    Collections.sort(models, Comparator { a: JSONObject, b: JSONObject -> a.getString("name").compareTo(b.getString("name")) } as java.util.Comparator<JSONObject>)
+    for (n in models) {
+        if (collectionTask.isCancelled()) {
+            Timber.e("doInBackgroundLoadModels :: Cancelled")
+            // onPostExecute not executed if cancelled. Return value not used.
+            return null
+        }
+        cardCount.add(col.models.useCount(n))
+    }
+    return Pair(models, cardCount)
 }
