+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionTask.kt
@@ -728,19 +728,7 @@ open class CollectionTask<Progress, Result>(val task: TaskDelegateBase<Progress,
      */
     class CountModels : TaskDelegate<Void, Pair<List<Model>, ArrayList<Int>>?>() {
         override fun task(col: Collection, collectionTask: ProgressSenderAndCancelListener<Void>): Pair<List<Model>, ArrayList<Int>>? {
-            Timber.d("doInBackgroundLoadModels")
-            val models = col.models.all()
-            val cardCount = ArrayList<Int>()
-            Collections.sort(models, Comparator { a: JSONObject, b: JSONObject -> a.getString("name").compareTo(b.getString("name")) } as Comparator<JSONObject>)
-            for (n in models) {
-                if (collectionTask.isCancelled()) {
-                    Timber.e("doInBackgroundLoadModels :: Cancelled")
-                    // onPostExecute not executed if cancelled. Return value not used.
-                    return null
-                }
-                cardCount.add(col.models.useCount(n))
-            }
-            return Pair(models, cardCount)
+            return countModels(col, collectionTask)
         }
     }
